+++ src/app/Models/Contact.php
@@ -1,0 +1,13 @@
+<?php
+namespace Mc388\SimpleCms\App\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+/**
+ * Class Contact
+ *
+ * @package Mc388\SimpleCms\App\Models
+ */
+class Contact extends Model
+{
+}