+++ src/database/migrations/2016_01_21_185824_create_contacts_table.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace Mc388\SimpleCms\Database\Migrations;
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+/**
+ * Class CreateContactsTable
+ */
+class CreateContactsTable extends Migration
+{
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('contacts', function (Blueprint $table) {
+            $table->increments('id');
+
+            $table->string('name');
+            $table->string('street');
+            $table->string('postal_code');
+            $table->string('city');
+            $table->string('phone')->nullable();
+            $table->string('mobile')->nullable();
+            $table->string('fax')->nullable();
+            $table->string('email')->nullable();
+
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::drop('contacts');
+    }
+
+}