+++ src/resources/views/site/contents/index.blade.php
@@ -1,0 +1,6 @@
+@if($contents->count() > 1)
+    @include('mc388-simple-cms::site.contents.root.multiple', ['contents' => $contents])
+@else
+    @include('mc388-simple-cms::site.contents.root.single', ['content' => $contents->first()])
+@endif
+