+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Create a new package
+- Add content model
+- Add contact model
+- Add sample data
+- Add content controller
+- Add frontend views
+- Add sitemap
+- Add gulp to generate css from scss
 
 ### Changed
 -